requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
qrcode[pil]==7.4.2
reportlab==4.0.8
websockets==12.0
aiofiles==23.2.1
pillow==10.1.0
